Added a new timer event system.
Instead of running everything in parallel it's now possible to have a
series of subtasks, each of which runs at a particular frequency. This
code maintains a linked list of event threads, sorted by timer ticks
(usually the tick is 1ms). Events at the queueu head are called when
their ticks run down to zero. The callbacks return the number of
milliseconds until the next invocation (or zero for one-offs).